backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
supabase==2.11.0
redis==5.2.1
httpx==0.28.1
# Newer Gemini 3.x models may need a recent SDK; upgrade if you see model-not-found errors.
google-generativeai>=0.8.4
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
